+++ tsdb/store.go
@@ -267,7 +267,7 @@ func (s *Store) Close() error {
 
 	// Close all the shards in parallel.
 	if err := s.walkShards(s.shardsSlice(), func(sh *Shard) error {
-		return sh.Close()
+		return sh.CloseFast()
 	}); err != nil {
 		return err
 	}
@@ -459,7 +459,7 @@ func (s *Store) DeleteDatabase(name string) error {
 			return nil
 		}
 
-		return sh.Close()
+		return sh.CloseFast()
 	}); err != nil {
 		return err
 	}